+++ index.go
@@ -39,23 +39,50 @@ func (c *Client) IndexPresent(table Table, index Index) error {
     }
 
     if (err != nil) {
-      c.Log(fmt.Sprintf("      * %s: failed to create", index.Name))
+      c.Log(fmt.Sprintf("      + %s ... create failed", index.Name))
       return err
     }
-    c.Log(fmt.Sprintf("      * %s: index created", index.Name))
-  } else {
-    c.Log(fmt.Sprintf("      * %s", index.Name))
+    c.clearTableIndexList(table)
   }
+  c.Log(fmt.Sprintf("      + %s", index.Name))
+  return nil
+}
+
+// IndexAbsent removes an index if it currently exists on the table
+func (c *Client) IndexAbsent(table Table, index Index) error {
+  indices, err := c.TableIndexList(table)
+  if (err != nil) { return err }
+
+  if (stringInSlice(index.Name, indices)) {
+    _, err = r.Table(table.Name).IndexDrop(index.Name).RunWrite(c.session)
+
+    if (err != nil) {
+      c.Log(fmt.Sprintf("      - %s ... drop failed", index.Name))
+      return err
+    }
+    c.clearTableIndexList(table)
+  }
+  c.Log(fmt.Sprintf("      - %s", index.Name))
 
   return nil
 }
 
 // TableIndexList returns a slice of index names on the table
 func (c *Client) TableIndexList(table Table) ([]string, error) {
+
+  if (len(c.indexListCache[table.Name]) > 0) {
+    return c.indexListCache[table.Name], nil
+  }
+
   res, err := r.Table(table.Name).IndexList().Run(c.session)
   if (err != nil) { return nil, err }
 
   indices := []string{}
   res.All(&indices)
-  return indices, nil
+  c.indexListCache[table.Name] = indices
+  return c.indexListCache[table.Name], nil
+}
+
+func (c *Client) clearTableIndexList(table Table) {
+  c.indexListCache[table.Name] = []string{}
 }