+++ index_test.go
@@ -35,6 +35,18 @@ func Test_IndexPresent_Simple_Create(t *testing.T) {
   expect(t, count, 1)
 }
 
+func Test_IndexPresent_Simple_IndexListFail(t *testing.T) {
+  table := Table{Name: "Test_IndexPresent"}
+  c, _ := NewClient(Connection{host, "non-existent-db"})
+
+  index := Index{
+    Name: "Test_IndexPresent_Simple_IndexListFail",
+  }
+  err := c.IndexPresent(table, index)
+  refute(t, err, nil)
+}
+
+
 func Test_IndexPresent_Simple_CreateFail(t *testing.T) {
   table := Table{Name: "Test_IndexPresent"}
   err := client.TablePresent(table)
@@ -117,4 +129,65 @@ func Test_IndexPresent_Exists(t *testing.T) {
   // Run it again!
   err = client.IndexPresent(table, index)
   expect(t, err, nil)
+}
+
+func Test_IndexAbsent_Success(t *testing.T) {
+  table := Table{Name: "Test_IndexAbsent_Success"}
+  err := client.TablePresent(table)
+  expect(t, err, nil)
+
+  index := Index{Name: "Test_IndexAbsent_Success"}
+  err = client.IndexPresent(table, index)
+  expect(t, err, nil)
+
+  indices, _ := client.TableIndexList(table)
+  expect(t, stringInSlice(index.Name, indices), true)
+
+  err = client.IndexAbsent(table, index)
+  expect(t, err, nil)
+
+  indices, _ = client.TableIndexList(table)
+  expect(t, stringInSlice(index.Name, indices), false)
+}
+
+func Test_IndexAbsent_IndexListFail(t *testing.T) {
+  table := Table{Name: "Test_IndexAbsent_IndexListFail"}
+  index := Index{Name: "Test_IndexAbsent_IndexListFail"}
+
+  badClient, _ := NewClient(Connection{host, "cheese"})
+  err := badClient.IndexAbsent(table, index)
+  refute(t, err, nil)
+}
+
+func Test_IndexAbsent_Fail(t *testing.T) {
+  table := Table{Name: "Test_IndexAbsent_Fail"}
+  index := Index{Name: "Test_IndexAbsent_Fail"}
+
+  badClient, _ := NewClient(Connection{host, "cheese"})
+  // Spoof the index list cache to think that the index already exists
+  badClient.indexListCache[table.Name] = []string{index.Name}
+  err := badClient.IndexAbsent(table, index)
+  refute(t, err, nil)
+}
+
+func Test_IndexAbsent_Already_Absent(t *testing.T) {
+  table := Table{Name: "Test_IndexAbsent_Fail"}
+  err := client.TablePresent(table)
+  expect(t, err, nil)
+
+  index := Index{Name: "Test_IndexAbsent_Fail"}
+  err = client.IndexPresent(table, index)
+  expect(t, err, nil)
+
+  indices, _ := client.TableIndexList(table)
+  expect(t, stringInSlice(index.Name, indices), true)
+
+
+  err = client.IndexAbsent(table, index)
+  expect(t, err, nil)
+
+  // Run it again!
+  err = client.IndexAbsent(table, index)
+  expect(t, err, nil)
+
 }