+++ table_test.go
@@ -3,6 +3,7 @@ package main
 
 import (
   "testing"
+  r "github.com/dancannon/gorethink"
 )
 
 func Test_TablePresent_Create(t *testing.T) {
@@ -13,6 +14,16 @@ func Test_TablePresent_Create(t *testing.T) {
 
   tables, _ := client.TableList()
   expect(t, stringInSlice(table.Name, tables), true)
+}
+
+func Test_TablePresent_TableList_Fail(t *testing.T) {
+  c, _ := NewClient(Connection{host, "non-existent-db"})
+  table := Table{
+    Name: "Test_TablePresent_Create_Fail####&",
+  }
+
+  err := c.TablePresent(table)
+  refute(t, err, nil)
 }
 
 func Test_TablePresent_Create_Fail(t *testing.T) {
@@ -32,21 +43,13 @@ func Test_TablePresent_Exists(t *testing.T) {
     Name: "Test_TablePresent_Exists",
   }
 
-  err := client.TablePresent(table)
+  _, err := r.Db(db).TableCreate(table.Name).RunWrite(client.session)
   expect(t, err, nil)
+  client.clearTableList()
 
-  tables, _ := client.TableList()
-  expect(t, stringInSlice(table.Name, tables), true)
-
-  // Look again, shuold not try to create the table
+  // Run it again!
   err = client.TablePresent(table)
   expect(t, err, nil)
-}
-
-func Test_TableList_Fail(t *testing.T) {
-  c, _ := NewClient(Connection{host, "non-existent-db"})
-  _, err := c.TableList()
-  refute(t, err, nil)
 }
 
 func Test_TableCreateOpts(t *testing.T) {
@@ -91,5 +94,46 @@ func Test_TableCreateOpts(t *testing.T) {
   expect(t, opts.PrimaryKey, nil)
   expect(t, opts.Durability, nil)
   expect(t, opts.DataCenter, nil)
+}
+
+func Test_TableAbsent_Success(t *testing.T) {
+  table := Table{Name: "Test_TableAbsent_Success"}
+
+  _, err := r.Db(db).TableCreate(table.Name).RunWrite(client.session)
+  expect(t, err, nil)
+  client.clearTableList()
+
+  err = client.TableAbsent(table)
+  expect(t, err, nil)
+
+  tables, _ := client.TableList()
+  expect(t, stringInSlice(table.Name, tables), false)
+}
+
+func Test_TableAbsent_TableListFail(t *testing.T) {
+  table := Table{Name: "Test_TableAbsent_TableListFail"}
+  badClient, _ := NewClient(Connection{host, "cheese"})
+  err := badClient.TableAbsent(table)
+  refute(t, err, nil)
+}
+
+func Test_TableAbsent_Fail(t *testing.T) {
+  table := Table{Name: "Test_TableAbsent_Fail"}
+  // Spoof the table list cache into thinking the table exists already
+  client.tableListCache = []string{table.Name}
+  err := client.TableAbsent(table)
+  refute(t, err, nil)
+}
+
+func Test_TableAbsent_Absent(t *testing.T) {
+  table := Table{Name: "Test_TableAbsent_Absent"}
+  _, err := r.Db(db).TableCreate(table.Name).RunWrite(client.session)
+  expect(t, err, nil)
+
+  err = client.TableAbsent(table)
+  expect(t, err, nil)
+
+  err = client.TableAbsent(table)
+  expect(t, err, nil)
 }