+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.0.1 - 2015-02-01
+### 0.1.0 - 2015-02-11
+
+* Remove tables from a database
+* Remove indices from a table
+
+### 0.0.1 - 2015-02-01
 
 * Intitial release!